requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.3.1
django-rest-framework-nested==0.0.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.5
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opencv-python-headless==4.2.0.32
pdf2image==1.14.0
Pillow==8.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pyPdf==1.13
PyPDF2==1.26.0
pytz==2020.5
pyzbar==0.1.8
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
